Halooglasi

+++ routines/routines.go
@@ -1,6 +1,8 @@
 package Routines
 
 import (
+	"errors"
+	"fmt"
 	"log"
 	"strconv"
 	Halooglasi "televito-parser/addsources/halooglasi"
@@ -112,7 +114,16 @@ func ReparseAllPages(class string) {
 			time.Sleep(2 * time.Minute)
 		}
 
-		if (len(adds)) == 0 {
+		if errors.Is(err, fmt.Errorf("page out of bounds")) {
+			page = 1
+			err = redisClient.DeleteKey(class + "_last_page")
+			if err != nil {
+				log.Println("Error deleting last page from redis: ", err)
+			}
+			continue
+		}
+
+		if (len(adds)) == 0 && class != "Halooglasi" {
 			page = 1
 
 			err = redisClient.DeleteKey(class + "_last_page")
